Working on improving the behaviour of the OutputStream objects - new implementation uses standard library stream objects, but uses XLE text conversion library - Fixed length stream and resizeable stream now share the same implementation

+++ Foreign/SonyLE/LevelEditorNativeRendering/NativeInterop/GameEngine.cs
@@ -304,7 +304,7 @@ namespace RenderingInterop
 
             unsafe
             {
-                ulong bufferSize = s_temporaryNativeBufferSize;
+                uint bufferSize = s_temporaryNativeBufferSize;
                 IntPtr pinnedPtr = s_temporaryNativeBuffer.AddrOfPinnedObject();
                 if (s_underlyingScene.GetProperty(
                     documentId, instanceId, typeId, propId,
